+++ docs/conf.py
@@ -20,6 +20,7 @@
 #
 import os
 import sys
+from typing import List
 
 sys.path.insert(0, os.path.abspath(".."))
 
@@ -49,7 +50,7 @@ master_doc = "index"
 
 # General information about the project.
 project = u"sxm-player"
-copyright = u"2018, Christopher Bailey"
+copyright = u"2018, Christopher Bailey"  # noqa: A001
 author = u"Christopher Bailey"
 
 # The version info for the project you're documenting, acts as replacement
@@ -85,7 +86,7 @@ todo_include_todos = False
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = "sphinx_rtd_theme"
+html_theme = "furo"
 
 # Theme options are theme-specific and customize the look and feel of a
 # theme further.  For a list of options available for each theme, see the
@@ -96,7 +97,7 @@ html_theme = "sphinx_rtd_theme"
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = []
+html_static_path: List[str] = []
 
 
 # -- Options for HTMLHelp output ---------------------------------------