+++ HISTORY.rst
@@ -2,6 +2,15 @@
 History
 =======
 
+0.2.0 (2021-08-15)
+------------------
+
+(WIP -- Pending)
+
+* Replaces requests with httpx
+* Updates linting
+* Replaces TravisCI with Github Actions
+
 0.1.0 (2018-12-25)
 ------------------